requirements.txt
amqp==2.4.2
aniso8601==3.0.2
asgiref==3.3.4
asn1crypto==1.4.0
astroid==2.3.0
billiard==3.5.0.5
cairocffi==1.2.0
CairoSVG==2.5.2
celery==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
configparser==5.0.1
cookies==2.2.1
cryptography==2.3.1
cssselect2==0.2.1
defusedxml==0.5.0
distlib==0.3.1
Django==3.0
django-allauth==0.37.1
django-autocomplete-light==3.4.1
django-background-tasks==1.2.0
django-bootstrap3==11.0.0
django-compat==1.0.15
django-cors-headers==3.2.1
django-crispy-forms==1.11.2
django-filter==2.0.0
django-report-builder==6.2.2
django-rest-auth==0.9.3
django-tables2==2.1.1
django-widget-tweaks==1.4.3
django-wkhtmltopdf==3.2.0
djangorestframework==3.12.4
et-xmlfile==1.0.1
filelock==3.0.12
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
html5lib==1.0.1
httpie==0.9.9
humanize==0.5.1
idna==2.6
isort==4.3.21
jdcal==1.4
kombu==4.2.2.post1
lazy-object-proxy==1.6.0
Markdown==2.6.9
mccabe==0.6.1
oauthlib==2.1.0
odfpy==1.3.6
openpyxl==2.4.9
pbr==5.5.0
Pillow==8.1.2
pipenv==2018.7.1
postgres==2.2.2
promise==2.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
psycopg2-pool==1.1
pycparser==2.19
Pygments==2.2.0
PyJWT==1.6.4
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
PyMySQL==0.9.2
Pyphen==0.9.5
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.7.5
python3-openid==3.1.0
pytz==2018.7
PyYAML==3.13
redis==3.0.1
requests==2.24.0
requests-oauthlib==1.0.0
requests-toolbelt==0.9.1
responses==0.9.0
rollbar==0.14.5
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.0
stevedore==3.2.2
tablib==0.12.1
text-unidecode==1.3
tinycss2==1.0.2
toml==0.10.2
typed-ast==1.4.0
typing==3.6.6
unicodecsv==0.14.1
urllib3==1.22
vine==5.0.0
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wcwidth==0.2.5
WeasyPrint==48
webencodings==0.5.1
wrapt==1.11.2
xlrd==1.1.0
xlwt==1.3.0
